chore 📦: Update dependencies in requirements files
This commit updates the `requirements.dev.txt` and `requirements.txt` files to ensure compatibility with the latest versions of dependencies. The changes include updating `pytest` and `pytest-cov` to their latest versions, `prompt-toolkit` to a more recent version, `rich` to a specific version, and `questionary` to a more recent version. Additionally, the `typing` dependency is updated to include `python_version < "3.7`. These changes ensure that the project uses the latest and most secure versions of its dependencies, improving stability and security.